Don't rebind VFs by default, run earlier
If a system is to be configured with the Mellanox PFs as part of a
bond to enable VF LAG, there must be no bound VFs at the time the
PFs are put into the bond.

We also need to run the service earlier at boot to allow
initialization to be complete prior to netplan applying runtime
network config which may include putting interfaces into bonds
etc.

https://community.mellanox.com/s/article/Configuring-VF-LAG-using-TC